WEA-27 Move realm forecast model to seperate package
