+++ app/src/main/java/com/ayush/weatherapp/realm/RealmAppModule.java
@@ -1,11 +1,11 @@
 package com.ayush.weatherapp.realm;
 
-import com.ayush.weatherapp.realm.model.CurrentForecast;
-import com.ayush.weatherapp.realm.model.DailyData;
-import com.ayush.weatherapp.realm.model.DailyForecast;
-import com.ayush.weatherapp.realm.model.Forecast;
-import com.ayush.weatherapp.realm.model.HourlyData;
-import com.ayush.weatherapp.realm.model.HourlyForecast;
+import com.ayush.weatherapp.realm.model.forecast.CurrentForecast;
+import com.ayush.weatherapp.realm.model.forecast.DailyData;
+import com.ayush.weatherapp.realm.model.forecast.DailyForecast;
+import com.ayush.weatherapp.realm.model.forecast.Forecast;
+import com.ayush.weatherapp.realm.model.forecast.HourlyData;
+import com.ayush.weatherapp.realm.model.forecast.HourlyForecast;
 import io.realm.annotations.RealmModule;
 
 @RealmModule(classes = {
