+++ app/src/main/java/com/ayush/weatherapp/repository/weather/LocalWeatherDataStoreImpl.java
@@ -1,8 +1,7 @@
 package com.ayush.weatherapp.repository.weather;
 
 import com.ayush.weatherapp.realm.RealmUtils;
-import com.ayush.weatherapp.realm.model.Forecast;
-import io.reactivex.Observable;
+import com.ayush.weatherapp.realm.model.forecast.Forecast;
 import io.reactivex.Single;
 
 public class LocalWeatherDataStoreImpl implements WeatherDataStore {
